requirements.txt
torch>=2.0.0
torch-geometric>=2.4.0
torch-scatter>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
rdkit>=2023.03.1
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0

